+++ packages/picasso.js/src/web/components/brush-area/index.js
@@ -1,0 +1,5 @@
+import definition from './brush-area';
+
+export default function areaBrush(picasso) {
+  picasso.component('brush-area', definition);
+}